Implement popover to preview items
